+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - fixed autocomplete for *.should methods
 - fixed browser.switch_to
 
+### known issues:
+- entity.with_ does not work in case of shared browser :(
+  - where entity = browser | element | collection
+
 
 ## 2.0.0a4 (released on 30.12.2019)
 - fixed default browser_name handling in shared config; implemented some old opts in config